--- src/main/java/com/korea/project/entity/ResourceEntity.java
@@ -1,69 +1,0 @@
-package com.korea.project.entity;
-
-import javax.persistence.*;
-import java.util.Objects;
-
-@Entity
-@Table(name = "resource", schema = "korea")
-public class ResourceEntity {
-    private Integer id;
-    private String code;
-    private String name;
-    private Integer parentid;
-
-    @Id
-    @Column(name = "id", nullable = false)
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    @Basic
-    @Column(name = "code", nullable = false, length = 45)
-    public String getCode() {
-        return code;
-    }
-
-    public void setCode(String code) {
-        this.code = code;
-    }
-
-    @Basic
-    @Column(name = "name", nullable = false, length = 45)
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    @Basic
-    @Column(name = "parentid", nullable = true)
-    public Integer getParentid() {
-        return parentid;
-    }
-
-    public void setParentid(Integer parentid) {
-        this.parentid = parentid;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        ResourceEntity that = (ResourceEntity) o;
-        return Objects.equals(id, that.id) &&
-                Objects.equals(code, that.code) &&
-                Objects.equals(name, that.name) &&
-                Objects.equals(parentid, that.parentid);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id, code, name, parentid);
-    }
-}
