+++ src/main/java/com/korea/project/entity/User.java
@@ -1,0 +1,126 @@
+package com.korea.project.entity;
+
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
+import javax.persistence.OneToMany;
+import javax.persistence.Table;
+import java.sql.Timestamp;
+import java.util.Collection;
+import java.util.Objects;
+
+@Entity
+@Table(name="user")
+public class User {
+    private Integer id;
+    private String firstname;
+    private Timestamp joindate;
+    private String lastname;
+    private String password;
+    private Integer role;
+    private String username;
+    private Collection<Project> projectsById;
+
+    @Id
+    @GeneratedValue(strategy=GenerationType.IDENTITY)
+    @Column(name = "id", nullable = false)
+    public Integer getId() {
+        return id;
+    }
+
+    public void setId(Integer id) {
+        this.id = id;
+    }
+
+
+    @Column(name = "firstname", nullable = false, length = 50)
+    public String getFirstname() {
+        return firstname;
+    }
+
+    public void setFirstname(String firstname) {
+        this.firstname = firstname;
+    }
+
+
+    @Column(name = "joindate", nullable = false)
+    public Timestamp getJoindate() {
+        return joindate;
+    }
+
+    public void setJoindate(Timestamp joindate) {
+        this.joindate = joindate;
+    }
+
+
+    @Column(name = "lastname", nullable = false, length = 50)
+    public String getLastname() {
+        return lastname;
+    }
+
+    public void setLastname(String lastname) {
+        this.lastname = lastname;
+    }
+
+
+    @Column(name = "password", nullable = false, length = 50)
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+
+
+    @Column(name = "role", nullable = true)
+    public Integer getRole() {
+        return role;
+    }
+
+    public void setRole(Integer role) {
+        this.role = role;
+    }
+
+
+    @Column(name = "username", nullable = false, length = 50)
+    public String getUsername() {
+        return username;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        User user = (User) o;
+        return Objects.equals(id, user.id) &&
+                Objects.equals(firstname, user.firstname) &&
+                Objects.equals(joindate, user.joindate) &&
+                Objects.equals(lastname, user.lastname) &&
+                Objects.equals(password, user.password) &&
+                Objects.equals(role, user.role) &&
+                Objects.equals(username, user.username);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(id, firstname, joindate, lastname, password, role, username);
+    }
+
+    @OneToMany(mappedBy = "userByUserid")
+    public Collection<Project> getProjectsById() {
+        return projectsById;
+    }
+
+    public void setProjectsById(Collection<Project> projectsById) {
+        this.projectsById = projectsById;
+    }
+}
