--- src/main/java/com/korea/project/entity/UserEntity.java
@@ -1,118 +1,0 @@
-package com.korea.project.entity;
-
-import javax.persistence.*;
-import java.sql.Timestamp;
-import java.util.Objects;
-
-@Entity
-@Table(name = "user", schema = "korea")
-public class UserEntity {
-    private Integer id;
-    private String username;
-    private String firstname;
-    private String lastname;
-    private String password;
-    private String salt;
-    private Timestamp joindate;
-    private Integer role;
-
-    @Id
-    @Column(name = "id", nullable = false)
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    @Basic
-    @Column(name = "username", nullable = false, length = 50)
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    @Basic
-    @Column(name = "firstname", nullable = false, length = 50)
-    public String getFirstname() {
-        return firstname;
-    }
-
-    public void setFirstname(String firstname) {
-        this.firstname = firstname;
-    }
-
-    @Basic
-    @Column(name = "lastname", nullable = false, length = 50)
-    public String getLastname() {
-        return lastname;
-    }
-
-    public void setLastname(String lastname) {
-        this.lastname = lastname;
-    }
-
-    @Basic
-    @Column(name = "password", nullable = false, length = 50)
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    @Basic
-    @Column(name = "salt", nullable = false, length = 50)
-    public String getSalt() {
-        return salt;
-    }
-
-    public void setSalt(String salt) {
-        this.salt = salt;
-    }
-
-    @Basic
-    @Column(name = "joindate", nullable = false)
-    public Timestamp getJoindate() {
-        return joindate;
-    }
-
-    public void setJoindate(Timestamp joindate) {
-        this.joindate = joindate;
-    }
-
-    @Basic
-    @Column(name = "role", nullable = true)
-    public Integer getRole() {
-        return role;
-    }
-
-    public void setRole(Integer role) {
-        this.role = role;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        UserEntity that = (UserEntity) o;
-        return Objects.equals(id, that.id) &&
-                Objects.equals(username, that.username) &&
-                Objects.equals(firstname, that.firstname) &&
-                Objects.equals(lastname, that.lastname) &&
-                Objects.equals(password, that.password) &&
-                Objects.equals(salt, that.salt) &&
-                Objects.equals(joindate, that.joindate) &&
-                Objects.equals(role, that.role);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id, username, firstname, lastname, password, salt, joindate, role);
-    }
-}
