+++ src/main/resources/application.properties
@@ -3,5 +3,10 @@ spring.datasource.username=springdbuser
 spring.datasource.password=Test@1234
 spring.datasource.driverClassName=com.mysql.cj.jdbc.Driver
 spring.jpa.show-sql=true
-spring.jpa.hibernate.dialect=org.hibernate.dialect.MySQLDialect
-spring.jpa.hibernate.ddl-auto = create-drop
+spring.jpa.hibernate.dialect=org.hibernate.dialect.MySQL5Dialect
+spring.jpa.hibernate.ddl-auto = update
+
+database.hibernate.schema_update = update
+database.hibernate.show_sql = false
+database.hibernate.format_sql = true
+database.hibernate.use_sql_comments = true
